SC-4075 add enable student migration field
